requirements.txt
numpy
pandas
opencv-python
mss
Pillow
pynput
pytesseract